+++ src/config.php
@@ -38,4 +38,10 @@ return [
 
     // The URI patterns to exclude from static file caching (overrides any matching patterns to include). The second variable represents a site ID, or a blank string for all sites.
     //'excludedUriPatterns' => [["contact",""]],
+
+    // Whether elements should be cached in the database.
+    //'cacheElements' => true,
+
+    // Whether element queries should be cached in the database.
+    //'cacheElementQueries' => true,
 ];